requirements.txt
asgiref==3.5.0
astroid==2.9.3
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
Django==4.0.2
djangorestframework==3.13.1
h11==0.13.0
idna==3.3
isort==5.10.1
jellyfish==0.9.0
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.1.0
networkx==2.6.3
numpy==1.22.0
outcome==1.1.0
pathspec==0.9.0
platformdirs==2.4.1
pycparser==2.21
pylint==2.12.2
pyOpenSSL==21.0.0
pytz==2021.3
regex==2021.11.10
requests==2.27.1
segtok==1.5.11
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
tabulate==0.8.9
toml==0.10.2
tomli==2.0.0
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.7
webdriver-manager==3.5.3
wrapt==1.13.3
wsproto==1.0.0
yake @ git+https://github.com/LIAAD/yake@96e4a233f94ab384de30d057a8b5c889a3a21ed5
